implement reprojection

+++ src/viewer/renderer/Texture.js
@@ -9,9 +9,9 @@ export class Texture {
 		this.texture = gl.createTexture();
 		gl.bindTexture(gl.TEXTURE_2D, this.texture);
 		if (immutable) {
-			gl.texStorage2D(gl.TEXTURE_2D, 1, gl.RGBA16F, width, height);
+			gl.texStorage2D(gl.TEXTURE_2D, 1, gl.RGBA32F, width, height);
 		} else {
-			gl.texImage2D(gl.TEXTURE_2D, 0, gl.RGBA16F, width, height, 0,  gl.RGBA, gl.FLOAT, null);
+			gl.texImage2D(gl.TEXTURE_2D, 0, gl.RGBA32F, width, height, 0,  gl.RGBA, gl.FLOAT, null);
 		}
 		gl.texParameteri(gl.TEXTURE_2D, gl.TEXTURE_MIN_FILTER, gl.LINEAR);
 		gl.texParameteri(gl.TEXTURE_2D, gl.TEXTURE_MAG_FILTER, gl.LINEAR);
